Refactor file package as a standard go package practice
